+++ src/metrology/lib/protocol.clj
@@ -281,6 +281,17 @@
                          (pprint m))))
                  coll)))))))
 
+(defn custom-html
+  ""
+  [coll]
+  (if (zero? (count coll))
+      (string/join
+        "\n"
+        (map (fn [m]
+               (li {:class "appendix-section"}
+                   (:html m)))
+             coll))))
+
 (defn page-2
   "Приложение к протоколу поверки."
   [m]
@@ -300,7 +311,8 @@
       (ol
         (when (:sw_version m)
               (sw-version m))
-        (measurements-table (:measurements m))))
+        (measurements-table (:measurements m))
+        (custom-html (:html m))))
     (page-footer 2 2)))
 
 (defn protocol