+++ CHANGELOG.md
@@ -3,13 +3,13 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased] - 2023-11-08
 ### Changed
+- Added custom html content in the data base and protocol funcitons of shell.
 - The report pages opening in the browser through `clojure.java.shell` lib.
 - Counteragents html-report.
 - Added some hash-tags to `shell.clj` source. More effectively navigation.
 
 ## [0.1.5]
 ### Changed
-- Add a new arity to `make-widget-async` to provide a different widget shape.
 
 ## [0.1.4] - 2023-10-18
 ### Fixed